requirements.txt
flask
gunicorn
opencv-python-headless
pillow
transformers
torch